Added uart tx to python, changed iRAM
iRAM no longer works with a full .mem file. instead, it is booted with a very simple HALT bootloader that allows you to transmit in a program at boot time. This makes each individual test slightly slower, but this is more compact than requiring an entire ROM bootloader, as well as an iRAM block, as well as a dedicated RAM block.